documentation

+++ mlprodict/tools/graphs.py
@@ -591,12 +591,11 @@ def onnx2bigraph(model_onnx, recursive=False, graph_type='basic'):
             from mlprodict.tools import get_opset_number_from_onnx
             from mlprodict.tools.graphs import onnx2bigraph
 
-            idi = numpy.identity(2)
+            idi = numpy.identity(2).astype(numpy.float32)
             opv = get_opset_number_from_onnx()
             A = OnnxAdd('X', idi, op_version=opv)
             B = OnnxSub(A, 'W', output_names=['Y'], op_version=opv)
-            onx = B.to_onnx({'X': idi.astype(numpy.float32),
-                             'W': idi.astype(numpy.float32)})
+            onx = B.to_onnx({'X': idi, 'W': idi})
             bigraph = onnx2bigraph(onx)
             graph = bigraph.display_structure()
             text = graph.to_text()
